+++ src/index.ts
@@ -1,4 +1,3 @@
 export * from './trait'
-export * from './callable'
 export * from './mutate'
 export * from './relational'
